MR : Feat: Add Product repository, Buy repository, Point repository | hong =>

See merge request s09-webmobile3-sub2/S09P12A304!38

+++ BackEnd/ggong/src/main/java/com/a304/ggong/dto/PointDto.java
@@ -9,7 +9,7 @@ import lombok.*;
 @ToString
 public class PointDto {
     //포인트 번호(id)
-    private int pointNo;
+    private Long pointNo;
     //포인트 적립/사용 시간
     private String eventTime;
     //적립/사용 포인트(적립은 +, 사용은 -)
